#33 Move companion build instances into their respective Rope's trait
